Refactor device redux reducers and actions
