+++ app/actions/device/index.js
@@ -1,2 +1,2 @@
-export * from './notifyDeviceUpdated'
 export * from './updateDeviceData'
+export * from 'routes/extension/Password/store/actions/notifyDeviceUpdated'
