+++ app/routes/extension/Password/containers/actions.js
@@ -1,0 +1,3 @@
+import { notifyDeviceUpdated } from '../store/actions'
+
+export default { notifyDeviceUpdated }
