+++ app/routes/extension/Password/store/actions/index.js
@@ -1,0 +1,1 @@
+export * from './notifyDeviceUpdated'
