+++ app/routes/extension/Password/store/selectors/index.js
@@ -1,0 +1,3 @@
+export const accountSelector = (state) => state.account
+
+export const deviceSelector = (state) => state.device
